% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI_cox.R
\name{KI_coxph}
\alias{KI_coxph}
\title{Confidence Interval for Hazard Ratio as approximation for the Ratio of Parameters for Two Exponentially Distributed Samples}
\usage{
KI_coxph(X, Y, Dx, Dy, alpha = 0.05)
}
\arguments{
\item{X}{numeric matrix -- Observed survival times in first group}

\item{Y}{numeric matrix -- Observed survival times in second group}

\item{Dx}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for first group}

\item{Dy}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for second group}

\item{alpha}{confidence level}
}
\value{
a data.frame with columns C1 (lower CI) and C2 (upper CI)
}
\description{
Confidence Interval for Hazard Ratio as approximation for the Ratio of Parameters for Two Exponentially Distributed Samples
}
