% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI_logrank.R
\name{KI_logrank}
\alias{KI_logrank}
\title{Pseudo (!) Confidence Interval for Logrank test for the Ratio of Parameters for Two Exponentially Distributed Samples}
\usage{
KI_logrank(X, Y, Dx, Dy, alpha = 0.05)
}
\arguments{
\item{X}{numeric matrix -- Observed survival times in first group}

\item{Y}{numeric matrix -- Observed survival times in second group}

\item{Dx}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for first group}

\item{Dy}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for second group}

\item{alpha}{confidence level}
}
\value{
a data.frame with columns L1 (lower CI) and L2 (upper CI), giving pseudo confidence levels.
}
\description{
Pseudo (!) Confidence Interval for Logrank test for the Ratio of Parameters for Two Exponentially Distributed Samples
}
\details{
this function just performs a logrank test and returns
         \itemize{
         \item (0, 1), if the test is significant and mean(Y) > mean(X)
         \item (1, Inf), if the test is significant and mean(X) > mean(Y)
         \item (0, Inf), if the test is not significant.
         }
         Warning! It shall be noted, that theese pseudo confidence intervalls do not hold any confidence levels as it may be intended.
         Rather, this procedure is implemented to fit a power simulation for the logrank test into the framwork of the other confidence intervals.
}
