% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI_survreg.R
\name{KI_survreg}
\alias{KI_survreg}
\title{Confidence Interval for Regression Parameter for the Ratio of Parameters for Two Exponentially Distributed Samples}
\usage{
KI_survreg(X, Y, Dx, Dy, alpha = 0.05)
}
\arguments{
\item{X}{numeric matrix -- Observed survival times in first group}

\item{Y}{numeric matrix -- Observed survival times in second group}

\item{Dx}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for first group}

\item{Dy}{numeric matrix -- Observation indicator (1 for events, 0 for censorings) for second group}

\item{alpha}{confidence level}
}
\value{
a data.frame with columns R1 (lower CI) and R2 (upper CI)
}
\description{
Confidence Interval for Regression Parameter for the Ratio of Parameters for Two Exponentially Distributed Samples
}
