% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KI_wald.R
\name{KI_wald}
\alias{KI_wald}
\title{Wald Confidence Interval for Ratio of Parameters for Two Exponentially Distributed Samples}
\usage{
KI_wald(x_mean, y_mean, nx, ny, dx, dy, alpha = 0.05)
}
\arguments{
\item{x_mean}{Mean of first group}

\item{y_mean}{Mean of second group}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{dx}{observation probability (=1-censoring prob.) of first group}

\item{dy}{observation probability (=1-censoring prob.) of second group}

\item{alpha}{confidence level}
}
\value{
a data.frame with columns W1 (lower CI) and W2 (upper CI)
}
\description{
Wald Confidence Interval for Ratio of Parameters for Two Exponentially Distributed Samples
}
\examples{
res <- KI_wald(x_mean = c(3,4),
               y_mean = c(4,6),
               nx     = c(100, 200),
               ny     = c(200, 100),
               dx     = c(.4, .5),
               dy     = c(.5, .7),
               alpha  = .05)

}
