% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_probability.R
\name{observation_probability}
\alias{observation_probability}
\title{Calculate Observation Probability given Distributions for Event and Censoring Times}
\usage{
observation_probability(dist_E, param_E = NULL, dist_C, param_C = NULL,
  N = 1e+07)
}
\arguments{
\item{dist_E}{distribution identifier for event distribution, see \code{\link{random_from_dist}}}

\item{param_E}{parameter set for event distribution, see \code{\link{random_from_dist}}}

\item{dist_C}{distribution identifier for censoring distribution, see \code{\link{random_from_dist}}}

\item{param_C}{parameter set for censoring distribution, see \code{\link{random_from_dist}}}

\item{N}{number of replicates}
}
\value{
a numeric value 0<p<1, denoting observation probability
}
\description{
Calculate Observation Probability given Distributions for Event and Censoring Times
}
\details{
For a given distribution of E and C, N samples each are drawn to
         approximate the probability for E_i < C_i, say the event happens before censoring
}
\examples{
  d_o <- observation_probability(dist_E  = "exp",
                                 param_E = c(1),
                                 dist_C  = "accrual-followup",
                                 param_C = c(.05, 5, 2),
                                 N       = 1e6)


}
