% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_generator.R
\name{random_from_dist}
\alias{random_from_dist}
\title{Generate Random Numbers from a specified distribution}
\usage{
random_from_dist(dist, param = NULL, N = 1e+07)
}
\arguments{
\item{dist}{a character specifying the desired distribution}

\item{param}{A Vector giving the parameters of the distributions respectively, see details}

\item{N}{Number of Randoms}
}
\value{
N random numbers
}
\description{
Generate Random Numbers from a specified distribution
}
\details{
This function is a wrapper for various random number generators
         (for different distributions). Its main purpose is to deliver
         an interface for mixed distributions.

         Supported are
         \itemize{
           \item exp (exponential distribution), param[1] = parameter
           \item weibull (weibull distribution), param[1] = shape, param[2] = scale
           \item norm (normal distribution), param[1] = mu, param[2] = sd, if negative set to 0
           \item unif (uniform distribution from 0 to param[1])
           \item gamma (gamma distribution with shape = param[1] and scale = param[2])
           \item dirac not really random, but N replicates of param[1]
           \item triangle (random numbers with parabolic distribution), param[1] = length (= max - min), param[2] = max
         }
}
