% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_gradient.R
\name{SampleSize_gradient}
\alias{SampleSize_gradient}
\title{Sample size calculation based on gradient statistics}
\usage{
SampleSize_gradient(k1, dx, dy, alpha = 0.05, beta = 0.2, k0 = 1,
  c = 1)
}
\arguments{
\item{k1}{estimated ratio of parameters}

\item{dx}{observation probability in first group}

\item{dy}{observation probability in second group}

\item{alpha}{confidence level (type 1 error)}

\item{beta}{1 - Power (type 2 error)}

\item{k0}{Effect under Null Hypothesis (usually equal to 1)}

\item{c}{ratio ny/nx [where n* is sample size of first/second group]}
}
\value{
List (3 elements) containing estimated required sample size (observations + censorings)
        \itemize{
        \item in first group (Nx)
        \item in second group (Ny)
        \item altogether (N)
        }
}
\description{
Sample size calculation based on gradient statistics
}
\examples{
SampleSize_gradient(2, .9, .9)

}
