% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_graphic.R
\name{SampleSize_graphic}
\alias{SampleSize_graphic}
\title{Graphical User Interface for Sample Size Calculation}
\usage{
SampleSize_graphic()
}
\value{
nothing. But while processing, a graphic shell appears, which does all the calculations implemented here
}
\description{
Graphical User Interface for Sample Size Calculation
}
\examples{
\dontrun{SampleSize_graphic()}

}
