% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_quadratic_equation.R
\name{solve_quadratic_equation}
\alias{solve_quadratic_equation}
\title{helper function: solve quadratic equations (suports vectorized form)}
\usage{
solve_quadratic_equation(a, b, c, current_task = "<undefined>")
}
\arguments{
\item{a}{numeric (or numeric vector) for parameter}

\item{b}{numeric (or numeric vector) for parameter}

\item{c}{numeric (or numeric vector) for parameter}

\item{current_task}{character-identifier to improve traceback, if equation has no zeros}
}
\value{
a two-element list with both zeros (or a vector of such)
}
\description{
helper function: solve quadratic equations (suports vectorized form)
}
\details{
This function solves quadratic equations of the form \deqn{a*x^2 + b*x + c = 0}. It also solves multiple equations at once, when the parameters are passed as vectors.
}
\examples{
a <- c(1,2,3)
b <- c(0,4,6)
c <- c(-1,-6,-15)
res <- solve_quadratic_equation(a, b, c)

}
